Massively improve claim lookup and counting performance
Closes #229

+++ src/main/java/the_fireplace/clans/commands/details/CommandList.java
@@ -75,7 +75,7 @@ public class CommandList extends ClanSubCommand {
 					break;
 				case "land":
 				case "claims":
-					clans.sort(Comparator.comparingInt(Clan::getClaimCount));
+					clans.sort(Comparator.comparingLong(Clan::getClaimCount));
 					for (Clan clan : clans)
 						listItems.add(TranslationUtil.getTranslation(sender, "commands.clan.list.listitem", clan.getClaimCount(), clan.getName(), clan.getDescription()).setStyle(TextStyles.GREEN));
 					break;
